refactor: put user type into auth type
