+++ app/_shared/types/auth.d.ts
@@ -1,5 +1,25 @@
 import type { FetcherResponse } from '@/_shared/utils/fetcher';
-import type { UserType } from '@/_shared/types/user';
+import type { UserType } from '@/_shared/types/auth';
+
+export declare namespace UserType {
+  type AuthUserResponse = {
+    email: string;
+    id: number;
+    login_type: number;
+    name: string;
+    picture_url: string;
+    role_id: number;
+    username: string;
+  };
+
+  type AuthUserContext = {
+    userID: number;
+    email: string;
+    name: string;
+    pictureUrl: string;
+    webinarEnabled: boolean;
+  };
+}
 
 export declare namespace AuthType {
   type AuthorizeResponse = FetcherResponse & {
@@ -23,12 +43,12 @@ export declare namespace AuthType {
 
   type CurrentAuthExternalResponse = FetcherResponse & {
     message: string;
-    data: UserType.AuthUserData;
+    data: UserType.AuthUserResponse;
   };
 
   type CurrentAuthInternalResponse = FetcherResponse & {
     message: string;
-    data: UserType.AuthUserData | null;
+    data: UserType.AuthUserResponse | null;
   };
 
   type SignOutResponse = FetcherResponse & {
