--- app/_shared/types/user.d.ts
@@ -1,11 +1,0 @@
-export declare namespace UserType {
-  type AuthUserData = {
-    email: string;
-    id: number;
-    login_type: number;
-    name: string;
-    picture_url: string;
-    role_id: number;
-    username: string;
-  };
-}
